bypass some `rmdir()` tests in Windows

+++ test/rmdir.js
@@ -25,22 +25,25 @@ test('rmdirSync() on missing directory', (t) => {
   })
 })
 
-const FILE_PATH = __filename
+if (process.platform.indexOf('win') !== 0) {
+  // https://github.com/nodejs/node/issues/8797
+  const FILE_PATH = __filename
 
-test.cb('rmdir() on file instead of directory', (t) => {
-  lib.rmdir(FILE_PATH, (err) => {
-    t.truthy(err)
-    t.end()
+  test.cb('rmdir() on file instead of directory', (t) => {
+    lib.rmdir(FILE_PATH, (err) => {
+      t.truthy(err)
+      t.end()
+    })
   })
-})
 
-test('rmdir() on file instead of directory', (t) => {
-  return lib.rmdir(FILE_PATH)
-    .then(() => t.fail('should reject'), () => t.pass('does reject'))
-})
-
-test('rmdirSync() on file instead of directory', (t) => {
-  t.throws(() => {
-    lib.rmdirSync(FILE_PATH)
+  test('rmdir() on file instead of directory', (t) => {
+    return lib.rmdir(FILE_PATH)
+      .then(() => t.fail('should reject'), () => t.pass('does reject'))
   })
-})
+
+  test('rmdirSync() on file instead of directory', (t) => {
+    t.throws(() => {
+      lib.rmdirSync(FILE_PATH)
+    })
+  })
+}
